file_sharing/requirements.txt
asgiref==3.11.1
Django==5.2.11
joblib==1.5.3
lxml==6.0.2
numpy==2.4.2
PyPDF2==3.0.1
python-docx==1.2.0
scikit-learn==1.8.0
scipy==1.17.1
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
